celery/requirements.txt
celery
google-cloud-storage
pillow-simd
redis
